requirements.txt

Django>=4.2,<5.3
djangorestframework>=3.14
djangorestframework-simplejwt>=4.8
django-cors-headers>=3.14
